+++ app/src/main/java/es/uva/mangostas/sharedplaylist/ServerActivity.java
@@ -1,7 +1,7 @@
 package es.uva.mangostas.sharedplaylist;
 
+import android.content.Intent;
 import android.media.MediaPlayer;
-import android.os.AsyncTask;
 import android.os.Bundle;
 import android.os.Handler;
 import android.support.v7.app.AppCompatActivity;
@@ -22,20 +22,11 @@ import com.getbase.floatingactionbutton.FloatingActionsMenu;
 import com.google.android.youtube.player.YouTubeInitializationResult;
 import com.google.android.youtube.player.YouTubePlayer;
 import com.google.android.youtube.player.YouTubePlayerFragment;
-import com.google.api.client.http.HttpRequest;
-import com.google.api.client.http.HttpRequestInitializer;
-import com.google.api.client.http.javanet.NetHttpTransport;
-import com.google.api.client.json.jackson2.JacksonFactory;
-import com.google.api.services.youtube.YouTube;
-import com.google.api.services.youtube.model.SearchListResponse;
 import com.quinny898.library.persistentsearch.SearchBox;
 import com.quinny898.library.persistentsearch.SearchResult;
 
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Iterator;
-import java.util.List;
-import java.util.concurrent.ExecutionException;
 
 import es.uva.mangostas.sharedplaylist.Model.ShpMediaObject;
 import es.uva.mangostas.sharedplaylist.Model.ShpSong;
@@ -53,8 +44,6 @@ public class ServerActivity extends AppCompatActivity implements YouTubePlayer.O
      * Define a global instance of a Youtube object, which will be used
      * to make YouTube Data API requests.
      */
-    private static YouTube youtube;
-    private static final long NUMBER_OF_VIDEOS_RETURNED = 5;
 
     private MediaPlayer myMediaPlayer;
     private Handler handler;
@@ -80,117 +69,11 @@ public class ServerActivity extends AppCompatActivity implements YouTubePlayer.O
         fab_yt.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                browseOnYoutubeByTerm();
+                showSearchbox();
             }
         });
 
-        search.setSearchListener(new SearchBox.SearchListener() {
-
-            @Override
-            public void onSearchOpened() {
-                fab.collapse();
-            }
-
-            @Override
-            public void onSearchClosed() {
-                closeSearch();
-            }
-
-            @Override
-            public void onSearchTermChanged(final String term) {
-                // React to the search term changing
-                // Called after it has updated results
-
-                Log.d("ytSearch", "Empiezo la busqueda");
-
-                try {
-                    SearchListResponse searchResponse = new AsyncTask<Void, Void, SearchListResponse>() {
-                        @Override
-                        protected SearchListResponse doInBackground(Void... voids) {
-
-                            YouTube youtube = new YouTube.Builder(new NetHttpTransport(), new JacksonFactory(),
-                                    new HttpRequestInitializer() {
-                                        public void initialize(HttpRequest request) throws IOException {
-                                        }
-                                    }).setApplicationName("ytst-150316")
-                                    .build();
-
-                            Log.d("testYT", "Inicializado el YouTube");
-
-                            try {
-                                YouTube.Search.List searchYt = youtube.search().list("id,snippet");
-                                searchYt.setKey(APIKEY);
-                                searchYt.setQ(term);
-                                searchYt.setType("video");
-                                searchYt.setFields("items(id/kind,id/videoId,snippet/title,snippet/thumbnails/default/url)");
-                                searchYt.setMaxResults(NUMBER_OF_VIDEOS_RETURNED);
-
-                                Log.d("testYT", "Metida la info al objeto de consulta");
-                                SearchListResponse searchResponse = searchYt.execute();
-                                Log.d("testYT", "He realizado la consulta con exito");
-
-                                /*List<com.google.api.services.youtube.model.SearchResult> searchResultList = searchResponse.getItems();
-                                Iterator<com.google.api.services.youtube.model.SearchResult> iteratorSearchResults = searchResultList.iterator();*/
-
-                                Log.d("testYT", "Procesada la petición. ¡Tengo la info!");
-
-                                return searchResponse;
-
-    /*                            while (iteratorSearchResults.hasNext()) {
-                                    com.google.api.services.youtube.model.SearchResult singleVideo = iteratorSearchResults.next();
-                                    //ResourceId rId = singleVideo.getId();
-
-                                    SearchResult result = new SearchResult(singleVideo.getSnippet().getTitle());
-                                    search.addSearchable(result);
-
-                                   // Log.d("testYT", "Titulo: " + singleVideo.getSnippet().getTitle());
-                                }*/
-                            } catch (IOException e) {
-                                Log.d("testYT", "Estoy tirando la IOException");
-                            }
-
-                            return null;
-                        }
-
-                    }.execute((Void) null).get();
-
-                    List<com.google.api.services.youtube.model.SearchResult> searchResultList = searchResponse.getItems();
-                    Iterator<com.google.api.services.youtube.model.SearchResult> iteratorSearchResults = searchResultList.iterator();
-
-                    while (iteratorSearchResults.hasNext()) {
-                        com.google.api.services.youtube.model.SearchResult singleVideo = iteratorSearchResults.next();
-                        //ResourceId rId = singleVideo.getId();
-
-                        SearchResult result = new SearchResult(singleVideo.getSnippet().getTitle());
-                        search.addSearchable(result);
-
-                        Log.d("testYT", "Titulo: " + singleVideo.getSnippet().getTitle());
-                    }
-
-                } catch (InterruptedException e) {
-                    e.printStackTrace();
-                } catch (ExecutionException e) {
-                    e.printStackTrace();
-                }
-            }
-
-            @Override
-            public void onSearch(final String searchTerm) {
-
-            }
-
-            @Override
-            public void onResultClick(SearchResult result) {
-                //React to result being clicked
-                Log.d("ytSearch", "Result");
-            }
-
-            @Override
-            public void onSearchCleared() {
-
-            }
-
-        });
+        setSearchBoxListeners();
 
         // Get ListView object from xml
         listView = (ListView) findViewById(R.id.listView);
@@ -256,8 +139,53 @@ public class ServerActivity extends AppCompatActivity implements YouTubePlayer.O
 
     }
 
-    private void browseOnYoutubeByTerm() {
+    private void showSearchbox () {
         search.revealFromMenuItem(R.id.action_yt, this);
+    }
+
+    private void setSearchBoxListeners() {
+
+        search.setSearchListener(new SearchBox.SearchListener() {
+
+            @Override
+            public void onSearchOpened() {
+                fab.collapse();
+            }
+
+            @Override
+            public void onSearchClosed() {
+                closeSearch();
+            }
+
+            @Override
+            public void onSearchTermChanged(final String term) {
+            }
+
+            @Override
+            public void onSearch(final String searchTerm) {
+                SearchResult result = new SearchResult(searchTerm);
+                search.addSearchable(result);
+                startYoutubeResultsActivity(searchTerm);
+            }
+
+            @Override
+            public void onResultClick(SearchResult result) {
+                //React to result being clicked
+                Log.d("ytSearch", "Result");
+            }
+
+            @Override
+            public void onSearchCleared() {
+
+            }
+
+        });
+    }
+
+    private void startYoutubeResultsActivity(final String searchTerm) {
+        Intent intent = new Intent(this, YoutubeResultsActivity.class);
+        intent.putExtra("term" ,searchTerm);
+        startActivity(intent);
     }
 
     protected void closeSearch() {
