+++ app/src/main/res/layout/server.xml
@@ -82,15 +82,14 @@
 
     <!-- https://github.com/Quinny898/PersistentSearch -->
     <com.quinny898.library.persistentsearch.SearchBox
-        android:layout_width="wrap_content"
+        android:layout_width="48dp"
         android:id="@+id/searchbox"
-        android:layout_height="wrap_content"
+        android:layout_height="48dp"
         android:visibility="gone"
-        android:layout_marginTop="-6dp"
-        android:layout_marginLeft="-6dp"
-        android:layout_marginRight="-6dp"
         android:layout_alignParentLeft="true"
-        android:layout_alignParentTop="true" />
+        android:layout_alignParentTop="true"
+        tools:layout_editor_absoluteY="0dp"
+        tools:layout_editor_absoluteX="0dp" />
 
     <ListView
         tools:layout_constraintRight_creator="1"
